anapath/frontend/requirements.txt

# img
pillow
opencv-python-headless

uvicorn
fastapi

python-multipart


# web
streamlit
requests
python-dotenv
